// src/app/shared/models/employeFilteringData.model.ts
export class EmployeFilteringData {
    nationalityId: string;
    professionId: string;
    housingId: string;
    iDNum: string;
    professionName: string;
    sex: number[];
    marital_Status: number[];
    age: string;
    experience: number[]
    //  AgeOperand 
    religion: number[]
    canCook: boolean;
    treatOld: boolean;
    treatChildren: boolean;
    speakArabic: boolean;
    speakEnglish: boolean;
    pageIndex: number;
    pageSize: number;
    automaticDriving: boolean;
    manualDriving: boolean;
    cleaning: boolean;
    employees: Employee[];
    // [Symbol.iterator]() { return this; }
}
export interface Employee {
    isSelected:boolean;
    employeeId: string;
    videoName: string;
    jobTitle: string;
    jobNumber: string;
    nationalityId: string;
    sex: string;
    marital_Status: string;
    birthdate: Date;
    age: number;
    skills: string;
    image: string;
    idNumber: string;
    canCook: boolean;
    treatOld: boolean;
    statusCode: number;
    treatChildren: boolean;
    speakEnglish: boolean;
    speakArabic: boolean;
    automaticDriving: boolean;
    manualDriving: boolean;
    cleaning: boolean;
    notes: string;
    professionId: string;
    priorContracts: IndividualContract[]
    candidateId: string;
    employeeType: number;
    englishName: string;
    arabicName: string;
    name: string;
    nationalityName: string;
    professionName: string;
    religion: number;
    religionName: string;
    extraNotes: string;
    gender: number;
    genderName: string;
    maritalStatus: number;
    maritalStatusName: string;
    vedioUrl: string;
    candidate: Candidate;
    prevExperiences: PrevExperience[]
    empExperience: number;
    experiancePeriod: number;
    housingBuildingId: string;
}
interface IndividualContract {
    ContactId: string;
    ContactName: string;
    ContractDate: Date;
    ContractEndDate: Date;
    ContractNumber: string;
    Address: string;
    ContactAddress: string;
    ContractTypeId: string;
    ContractTypeName: string;
    FinalPrice: number;
    Insurance: number;
    FloorNo: number;
    HouseNo: string;
    HouseType: number;
    IndividualContractId: string;
    IsPaid: boolean;
    WantToRenew: boolean;
    Latitude: string;
    Longitude: string;
    MapUrl: string;
    PartmentNumber: string;
    PricingId: string;
    PricingName: string;
    ProfessionId: string;
    ProfessionName: string;
    NationalityId: string;
    NationalityName: string;
    TotalAmount: string;
    TotalAmountWithVat: number;
    VatAmount: number;
    VatRate: number;
    AcceptTerms: boolean;
    PromotionCode: string;
    FromSite: boolean;
    Contact: Contact;
    RecaptchaStatus: boolean;
    CurrentStep: ContractStepsEnum;
    StepType: StepTypeEnum;
    // DynamicTemplateParts Template 
    MobileNumber: string;
    Notes: string;
    EmployeeId: string;
    NewEmployeeId: string;
    EmployeeName: string;
    Signature: string;
    CustomerNationalityName: string;
    CreatedOn: Date;
    IsProfileCompleted: boolean;
    ContractStopDays: number;
    ServiceEndDate: Date
    IqamaAttachmentId: string;
    FamilyCardId: string;
    BusinessCardId: string;
    CarInsuranceId: string;
    CarLicenseId: string;
    StatusCode: number;
    StatusCodeName: string;
    RenewedFrom: string;
    PriceType: number;
    RenewedToContract: string;
    MTHContractNo: string;
    RenewContractAttachment: string;
    Template: DynamicTemplateParts;
    Pricing: IndividualPricing
}
// ...
interface Candidate {
    name: string;
    nationalityName: string;
    professionName: string;
    religion: number;
    religionName: string;
    gender: number;
    genderName: string;
    maritalStatus: number;
    maritalStatusName: string;
    birthdate: Date;
    age: number;
    experiancePeriod: number;
    careOfElderly: boolean;
    canCook: boolean;
    speakEnglish: boolean;
    speakArabic: boolean;
    dealWithChildren: boolean;
    candidateId: string;
    washingClothes: boolean;
    ironing: boolean;
    cleaning: boolean;
    otherSkills: string;
    automaticDriving: boolean;
    manualDriving: boolean;
    clean: boolean;
}

interface PrevExperience {
    ProfessionName: string;
    CityName: string;
    FromDate: Date;
    ToDate: Date;
}